Rename output 'params' field so it works with latest Hugo
Latest Hugo gives an error:

Error: error building site: assemble: unable to cast []interface {}{"bmp"} of type []interface {} to map[string]interface{}

I think this is something about it interpreting .params as similar to .Params, a method that returns a map.

Anyway, let's call this 'format' instead, as these are output formats in the 'dot --help'.